services/preprocessing-service/requirements.txt
boto3==1.35.39
confluent-kafka==2.6.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
requests==2.32.3
orjson==3.10.7
Pillow==10.4.0
imagehash==4.3.1
opencv-python-headless==4.10.0.84
numpy==2.1.1
insightface==0.7.3
onnxruntime==1.18.1
